Allow syscalls to have return values
intr_ctx structures passed to exception handling routines need to be marked as
volatile otherwise gcc assumes that any modifications made to them may be
eliminated. This is not the case when for example we want to write to a register
to indicate the return value of a syscall.

Also, allow kthread routines to return an exit code and actually call do_exit()
when they are finished.