Demand load the /sbin/init binary, works perfectly except for the #GPF
Except for a GPF when restoring control to the program after having loaded
it. Not sure why, we're doing iret to the right location, the page tables
are setup correctly (if they weren't it'd generate a pagefault anyway) and
the binary is loaded in memory correctly. Verified that with pmemsave from
KVM and sha1sum.

Could it be another one of those lingering historical bugs?