First cut of pagecache.
Demand loading binaries works on the initial hard page fault, a subsequent soft
page fault (cache hit) but on a third access to the cached page it causes an
infinite loop constantly faulting the page in (seemingly OK) but then
restarting the faulting instruction generates another fault leading to the
loop. Argh, has man-flu got the better of my brain?