ide: introduce is_ata() function
Recognize drive type without requiring a possibly infinite loop, just retry 12
times. Also insert an I/O pause so as not to rape the device while waiting for
the busy flag to clear. If it really takes so long, can't we wait for the
device to generate an interrupt?

Anyway, now my disk and CD-ROM are both correctly recognized.